FiveLiner/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
httpx==0.27.2
openai==1.50.2
youtube-transcript-api==0.6.2
python-dotenv==1.0.1
pydantic==2.9.2
yt-dlp==2024.8.6

